refactor: Improve modularization and initialization for production readiness
